The previous approach was actually prone to ID collisions. There was at
least one Travis run in which the test
tracing/tracingUtil/generate random IDs failed with:

1) tracing/tracingUtil
     generate random IDs
       must generate random span IDs:
    AssertionError: found a non-unique ID at indices 13141 and 14324: 000625e2b97fffff === 000625e2b97fffff
    + expected - actual
    -000625e2b97fffff
    +an ID != 000625e2b97fffff

In addition, the previous approach only produced 12-13 digits, the
left-most characters where always zeroes. This issue would become even
more pronounced once we switch to 32 char trace IDs.

The new approach (inspired by
https://github.com/sindresorhus/crypto-random-string) uses Node.js'
crypto.randomBytes function instead of Math.random to generate IDs with
a lot lower probability for collisions.

It is a bit more expensive. A micro benchmark produced the following results:

Previous approach (Math.random):
- Generating 10000000 IDs with 16 characters took  9807 milliseconds.
- Generating 10000000 IDs with 32 characters took  9804 milliseconds.
- On average, it costs approximately 0.001 ms (1 μs) per ID.

New approach (crypto.randomBytes):
- Generating 10000000 IDs with 16 characters took 29985 milliseconds.
- Generating 10000000 IDs with 32 characters took 30601 milliseconds.
- On average, it costs approximately 0.003 ms (3 μs) per ID.

So it is slower by factor three, but we get much better and actually
uniqe IDs. We generate at most two IDs for each incoming request
(entry span ID and optionally a trace ID if there is no incoming
trace ID) and one per exit span. 3 μs is still absolutely negligible
when compared to processing an incoming HTTP request/
GRPc message/Kafka/RabbitMQ/... message or any exit span (DB request,
outgoing HTTP request, ...).